+++ src/lib/audio/episode-details.ts
@@ -1,5 +1,4 @@
-import type { EpisodeDetails } from '$lib/types';
 import { writable } from 'svelte/store';
+import type { EpisodeDetails } from '../types';
 
 export const episode_details = writable<EpisodeDetails | null>(null);
-export type EpisodeDetailsStore = typeof episode_details;