+++ src/lib/audio/episode-progress.ts
@@ -1,7 +1,7 @@
-import { audio_element, type AudioElementStore } from '$lib/audio/audio-element';
-import type { EpisodeProgress } from '$lib/types';
-import { secondsToTimestamp } from '$lib/utility';
-import { derived } from 'svelte/store';
+import { derived, type Readable } from 'svelte/store';
+import type { EpisodeProgress } from '../types';
+import { secondsToTimestamp } from '../utility';
+import { audio_element } from './audio-element';
 
 const default_episode_progress = {
 	current_time: 0,
@@ -9,21 +9,20 @@ const default_episode_progress = {
 	has_ended: false,
 } satisfies EpisodeProgress;
 
-export const episode_progress = derived<AudioElementStore, EpisodeProgress>(
-	audio_element,
-	($audio, set) => {
-		if (!$audio) return set(default_episode_progress);
+type ProgressStore = Readable<EpisodeProgress>;
 
-		function set_value() {
-			if (!$audio) return;
-			set({
-				current_time: $audio.currentTime,
-				timestamp: secondsToTimestamp($audio.currentTime),
-				has_ended: $audio.ended,
-			});
-		}
+export const episode_progress = derived(audio_element, ($audio, set) => {
+	if (!$audio) return set(default_episode_progress);
 
-		$audio.addEventListener('timeupdate', () => set_value());
-		return () => $audio.removeEventListener('timeupdate', () => set_value());
-	},
-);
+	function set_value() {
+		if (!$audio) return;
+		set({
+			current_time: $audio.currentTime,
+			timestamp: secondsToTimestamp($audio.currentTime),
+			has_ended: $audio.ended,
+		});
+	}
+
+	$audio.addEventListener('timeupdate', () => set_value());
+	return () => $audio.removeEventListener('timeupdate', () => set_value());
+}) satisfies ProgressStore;