+++ src/lib/types.ts
@@ -6,7 +6,7 @@ export type EpisodeProgress = { current_time: number; timestamp: string; has_end
 
 export type EpisodeAttributes = {
 	will_autoplay: boolean;
-	is_muted: boolean;
+	is_paused: boolean;
 	duration: number;
 	src: string;
 	start_at: number;
@@ -22,3 +22,7 @@ export interface UserPreferences {
 	playback_rate: number;
 	volume: number;
 }
+
+export type UserProgress = {
+	[key: string]: number;
+};