+++ src/lib/utility/index.ts
@@ -1,2 +1,2 @@
+export * from './get-pathname-from-url';
 export * from './seconds-to-timestamp';
-export * from './use-state';